+++ minitalk.h
@@ -20,8 +20,6 @@
 # include <signal.h>
 # include <unistd.h>
 # include <stdlib.h>
-// #include <string.h>
-// #include <stdlib.h>
 
 // so vscode will leave me alone
 # ifndef SIGUSR1
@@ -32,10 +30,10 @@
 # endif
 
 # ifndef BUFFER
-# 	define BUFFER 10
+#  define BUFFER 10
 # endif
 # ifndef BAUD
-#  define BAUD 150
+#  define BAUD 50
 # endif
 
 void	ft_putchar_fd(char c, int fd);
@@ -44,7 +42,11 @@ void	ft_putnbr_fd(int n, int fd);
 int		ft_atoi(const char *str);
 char	*ft_strcpy(char *dest, char *src);
 
-//utils
+// server utils
+void	send_confirmation(unsigned int *pid);
+int		calc_len(int count);
+
+// utils
 int		message_exit(int sig, int fd, char *msg);
 void	ft_quick_itoa(unsigned int n);
