+++ server.c
@@ -10,50 +10,16 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "minitalk.h"
 
-# include "minitalk.h"
+char	g_message[BUFFER];
 
-char	message[BUFFER];
-
-void	print_bits(unsigned char octet)
-{
-	for (int i = 7; i >= 0; i--)
-	{
-		if (octet & (1U << i))
-			write(1, "1", 1);
-		else
-			write(1, "0", 1);
-	}
-}
-
-void	send_confirmation(unsigned int *pid)
-{
-	write(1, "\n", 1);
-	ft_putstr_fd("END acknowledging:", 1);
-	ft_quick_itoa(*pid);
-	write(1, "\n", 1);
-	kill(*pid, SIGUSR1);
-	*pid = 0;
-}
-
-void write_reset(int len, unsigned int *count)
+void	write_reset(int len, unsigned int *count)
 {
 	len += 1;
-	// write(1, "len:", 5);
-	// ft_quick_itoa(len);
-	// ft_putstr_fd("-writitng-<", 1);
-	write(1, message, len);
-	// write(1, ">\n", 2);
+	write(1, g_message, len);
 	*count = 32;
-	message[0] = '\0';
-}
-
-int	calc_len(int count)
-{
-	if (count <= 32)
-		return (0);
-	count -= 31;
-	return (count / 8 - 1);
+	g_message[0] = '\0';
 }
 
 int	get_octet(int sig, unsigned int count, unsigned int *ret)
@@ -65,13 +31,6 @@ int	get_octet(int sig, unsigned int count, unsigned int *ret)
 		octet = octet | 1U;
 	if (count % 8 == 0)
 	{
-		// ft_quick_itoa(count);
-		// write(1, " : ", 3);
-		// print_bits(octet);
-		// write(1, &octet, 1);
-		// write(1, " : ", 3);
-		// ft_quick_itoa(calc_len(count));
-		// write(1, "\n", 1);
 		*ret = octet;
 		octet = 0;
 		return (0);
@@ -89,23 +48,23 @@ void	decoder(int sig, unsigned int *count, unsigned int *pid)
 		*pid = (*pid << 8 | octet);
 	else
 	{
-		message[calc_len(*count)] = octet;
-		// message[calc_len(*count) + 1] = '\0';
+		g_message[calc_len(*count)] = octet;
 		if (octet == 0)
 		{
 			write_reset(calc_len(*count), count);
 			send_confirmation(pid);
-			* count = 0;
+			*count = 0;
 		}
 		else if (calc_len(*count) == BUFFER)
 			write_reset(calc_len(*count), count);
 	}
 }
 
-void handeler(int sig)
+void	handeler(int sig)
 {
 	static unsigned int	count = 0;
 	static unsigned int	pid = 0;
+
 	count += 1;
 	if (sig == SIGUSR1)
 		decoder(1, &count, &pid);
@@ -114,7 +73,7 @@ void handeler(int sig)
 	if (sig == SIGTERM || sig == SIGINT)
 	{
 		ft_putstr_fd("\nTERMINATION\nbuffer_flush{", 1);
-		ft_putstr_fd(message, 1);
+		ft_putstr_fd(g_message, 1);
 		ft_putstr_fd("}flush_complete .. goodbye\n", 1);
 		exit(0);
 	}
@@ -127,21 +86,9 @@ int	main(void)
 	signal(SIGTERM, handeler);
 	signal(SIGINT, handeler);
 	ft_putstr_fd("Server pid: ", 1);
-	ft_putnbr_fd(getpid(), 1);
+	ft_quick_itoa(getpid());
 	write(1, "\n", 1);
 	while (1)
 		pause();
 	return (0);
 }
-
-// 01245689acdeghilklnop
-// 01245689acdeghilklnop
-// 0124
-//     5689
-//         acde
-//             f
-//             ghil
-//                 j
-//                   m
-//                 klno
-//                     p
